+++ apps/backend/src/characters/characters.module.ts
@@ -1,11 +1,12 @@
 import { Module } from '@nestjs/common';
 import { DatabaseModule } from '../database/database.module';
 import { ImagesModule } from '../images/images.module';
+import { TagsModule } from '../tags/tags.module';
 import { CharactersService } from './characters.service';
 import { CharactersResolver } from './characters.resolver';
 
 @Module({
-  imports: [DatabaseModule, ImagesModule],
+  imports: [DatabaseModule, ImagesModule, TagsModule],
   providers: [CharactersService, CharactersResolver],
   exports: [CharactersService],
 })